# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Brand OS API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: user
    description: User data operations
  - name: drafts
    description: Draft content operations
  - name: ai
    description: AI generation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out (clears session cookie)
      responses:
        "204":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /user/preferences:
    get:
      operationId: getPreferences
      tags: [user]
      summary: Get the current user's brand preferences
      responses:
        "200":
          description: Preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferencesResponse"
    put:
      operationId: updatePreferences
      tags: [user]
      summary: Update the current user's brand preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePreferencesBody"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferencesResponse"

  /user/suggestions:
    get:
      operationId: getSuggestions
      tags: [user]
      summary: Get personalised content suggestions
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SuggestionItem"

  /ai/structure:
    post:
      operationId: structureIdea
      tags: [ai]
      summary: Extract structured breakdown from a raw thought
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StructureIdeaBody"
      responses:
        "200":
          description: Structured breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StructuredBreakdown"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: AI error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai/generate:
    post:
      operationId: generateContent
      tags: [ai]
      summary: Generate LinkedIn content from confirmed structure
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateContentBody"
      responses:
        "200":
          description: Generated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedContent"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: AI error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai/refine:
    post:
      operationId: refineContent
      tags: [ai]
      summary: Refine existing content based on an instruction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefineContentBody"
      responses:
        "200":
          description: Refined content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefinedContent"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: AI error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /drafts:
    get:
      operationId: listDrafts
      tags: [drafts]
      summary: List all drafts for the current user
      responses:
        "200":
          description: List of drafts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Draft"
    post:
      operationId: createDraft
      tags: [drafts]
      summary: Create a new draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDraftBody"
      responses:
        "201":
          description: Created draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"

  /drafts/{id}:
    get:
      operationId: getDraft
      tags: [drafts]
      summary: Get a single draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateDraft
      tags: [drafts]
      summary: Update a draft's content or status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDraftBody"
      responses:
        "200":
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteDraft
      tags: [drafts]
      summary: Delete a draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        displayName:
          type: string
      required:
        - email
        - password
        - displayName

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        displayName:
          type: string
      required:
        - id
        - email
        - displayName

    PreferencesResponse:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        objective:
          type: string
        persona:
          type: string
        tone:
          type: string
        brandRole:
          type: string
        brandAudience:
          type: string
        brandBelief:
          type: string
        onboarded:
          type: boolean
      required:
        - id
        - userId
        - objective
        - persona
        - tone
        - brandRole
        - brandAudience
        - brandBelief
        - onboarded

    UpdatePreferencesBody:
      type: object
      properties:
        objective:
          type: string
        persona:
          type: string
        tone:
          type: string
        brandRole:
          type: string
        brandAudience:
          type: string
        brandBelief:
          type: string
        onboarded:
          type: boolean

    SuggestionItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        message:
          type: string
        action:
          type: string
      required:
        - id
        - type
        - message
        - action

    StructureIdeaBody:
      type: object
      properties:
        rawInput:
          type: string
        objective:
          type: string
        persona:
          type: string
        tone:
          type: string
      required:
        - rawInput
        - objective
        - persona
        - tone

    StructuredBreakdown:
      type: object
      properties:
        topic:
          type: string
        angle:
          type: string
        coreMessage:
          type: string
        whyItMatters:
          type: string
        hooks:
          type: array
          items:
            type: string
        narrativeFlow:
          type: array
          items:
            type: string
      required:
        - topic
        - angle
        - coreMessage
        - whyItMatters
        - hooks
        - narrativeFlow

    CarouselSlide:
      type: object
      properties:
        slide:
          type: integer
        title:
          type: string
        description:
          type: string
      required:
        - slide
        - title
        - description

    GenerateContentBody:
      type: object
      properties:
        rawInput:
          type: string
        objective:
          type: string
        persona:
          type: string
        tone:
          type: string
        structure:
          $ref: "#/components/schemas/StructuredBreakdown"
        selectedHook:
          type: string
      required:
        - rawInput
        - objective
        - persona
        - tone
        - structure
        - selectedHook

    GeneratedContent:
      type: object
      properties:
        post:
          type: string
        carousel:
          type: array
          items:
            $ref: "#/components/schemas/CarouselSlide"
        visual:
          type: string
      required:
        - post
        - carousel
        - visual

    RefineContentBody:
      type: object
      properties:
        content:
          type: string
        instruction:
          type: string
        tab:
          type: string
          enum: [post, carousel, visual]
      required:
        - content
        - instruction
        - tab

    RefinedContent:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    Draft:
      type: object
      properties:
        id:
          type: integer
        rawInput:
          type: string
        objective:
          type: string
        persona:
          type: string
        tone:
          type: string
        structuredBreakdown:
          $ref: "#/components/schemas/StructuredBreakdown"
        postOutput:
          type: ["string", "null"]
        carouselOutput:
          type: ["string", "null"]
        visualOutput:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, ready, published]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - rawInput
        - objective
        - persona
        - tone
        - structuredBreakdown
        - status
        - createdAt
        - updatedAt

    CreateDraftBody:
      type: object
      properties:
        rawInput:
          type: string
        objective:
          type: string
        persona:
          type: string
        tone:
          type: string
        structuredBreakdown:
          $ref: "#/components/schemas/StructuredBreakdown"
        postOutput:
          type: ["string", "null"]
        carouselOutput:
          type: ["string", "null"]
        visualOutput:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, ready, published]
      required:
        - rawInput
        - objective
        - persona
        - tone
        - structuredBreakdown
        - status

    UpdateDraftBody:
      type: object
      properties:
        postOutput:
          type: ["string", "null"]
        carouselOutput:
          type: ["string", "null"]
        visualOutput:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, ready, published]