# lib/api-zod/src/generated/types/registerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Brand OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterBody {
  email: string;
  /** @minLength 8 */
  password: string;
  displayName: string;
}