# lib/api-zod/src/generated/types/suggestionItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Brand OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SuggestionItem {
  id: string;
  type: string;
  message: string;
  action: string;
}