# lib/api-zod/src/generated/types/updatePreferencesBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Brand OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePreferencesBody {
  objective?: string;
  persona?: string;
  tone?: string;
  brandRole?: string;
  brandAudience?: string;
  brandBelief?: string;
  onboarded?: boolean;
}