# lib/api-zod/src/generated/types/userResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Brand OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserResponse {
  id: number;
  email: string;
  displayName: string;
}